fix variable allocation overriding offset if var is already param

+++ 03-code/headers/stddef.h
@@ -3,7 +3,7 @@
 
 #define NULL ((void*)0)
 
-typedef unsigned long size_t;
+typedef int size_t;
 typedef int ptrdiff_t;
 typedef int wchar_t;
